pyproject.toml
[project]
name = "agent-cookbooks"
version = "0.1.0"
description = "A collection of practical examples demonstrating various agentic AI patterns and techniques using LangChain, LangGraph, and related technologies"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=0.5.0",
    "deepeval>=3.1.8",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.0",
    "fastmcp>=2.10.6",
    "guardrails-ai>=0.6.7",
    "httpx[socks]>=0.27.2",
    "ipython>=9.6.0",
    "lancedb>=0.21.2",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.35",
    "langchain-qdrant>=0.2.1",
    "langchain-tavily>=0.2.12",
    "langextract>=1.0.9",
    "langgraph>=0.6.10",
    "langgraph-api>=0.4.46",
    "langgraph-bigtool>=0.0.3",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-cli[inmem]>=0.4.4",
    "langgraph-supervisor>=0.0.29",
    "litellm>=1.78.5",
    "mcp[cli]>=1.6.0",
    "numexpr>=2.10.2",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.7",
    "pylance>=0.25.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.40",
    "sqlite-vec>=0.1.6",
    "tantivy>=0.22.2",
    "yfinance>=0.2.55",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["examples*"]

[tool.uv]
package = true
